Split out insta tests
